final project

+++ src/main/java/com/task/bookmark/dto/FolderBookmarkDTO.java
@@ -3,9 +3,9 @@ package com.task.bookmark.dto;
 import jakarta.validation.constraints.NotBlank;
 
 public record FolderBookmarkDTO(
-        Integer bookmarkId,
+        Long bookmarkId,
         @NotBlank(message = "Title cannot be empty") String title,
         @NotBlank(message = "Url cannot be empty") String url,
-        Integer folderId
+        Long folderId
 ) {
 }
